Put informative labels on shell steps
Many sh or bat pipeline steps have an uninformative label. Groovy uses
the script as a label by default, but this is uninformative when the
script is something like "./steps.sh", and it's unreadable or
truncated when the script is very long. In such cases, use an
informative label, containing the job name.

In particular, this makes it possible to quickly search for the
execution of a given all.sh component in the Jenkins pipeline view:
search for "steps.sh #${job_name}", e.g.
"steps.sh #all_u16-test_clang_opt".