Pass branch information explicitly in an object
Some branch information was stored in fields of the `common` class. This
is fragile: in a given run, each module that uses `common` gets its
own instance of the `common` class. So when the `gen_jobs` module's
instance fills fields of `common`, those are not seen in methods of
modules other than `gen_jobs`.

To fix this, move the mutable state into a new object `BranchInfo
info` which gets passed around explicitly.

This commit will need follow-ups:

* `common.docker_tags` is mutable. It is currently used in a way that
  just happens to work, perhaps because the field is declared
  `static`, but it should presumably be explicitly passed around like
  the branch information.
* Currently, `common.get_branch_information` runs twice. After this
  commit it should no longer be necessary: the `info` structure can be
  passed around more.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>