Put the long unstash sequence into its own stage
This improves reporting in BlueOcean, which has a limit of 100 steps.
The outcome stage is >100 boring unstash steps followed by an
interesting step (running analyze_outcomes.py). Due to this long
sequence, the interesting step was not visible on BlueOcean. Fix this
by making unstashes a separate nested stage.

Fix #127.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>