Fix result analysis reporting
Unify and fix the way result analysis runs in pull request testing and
release testing.

Before, the logic to report failures to GitHub or
over email only covered the main test stage and not the result
analysis stage. Now a failure of result analysis will be reported like
a failure of a main test component. Fixes #120.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>