Make sure we have Python >=3.8 on all Docker images
Mbed TLS 3.6 LTS officially requires Python >=3.8, although we're
still testing with (and therefore supporting) Python 3.6. At some
point we'll switch our testing to 3.8. In preparation, make sure that
Python >=3.8 is available on all Docker images.

On Ubuntu 16.04, the most recent system Python is 3.5, and we install
more recent version (3.6 and now also 3.8) with pyenv.

On Ubuntu 18.04, `/usr/bin/python3` is 3.6 but 3.8 is available as a
package, so install it.

On Ubuntu 20.04, `/usr/bin/python3` is 3.8.

Note that this commit merely makes Python 3.8 available. It does not
change which Python version gets used in CI runs.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>